ADD: Tester finish with diff macoslinux control

+++ tester/tester.sh
@@ -20,6 +20,7 @@ printf "\e[0;1;94m\
 #                      ____) | | |  | | |_| \__ \    / / | || |                  #
 #                     |_____/|_|_|  |_|\__,_|___/   /_/  |_||_|                  #
 #                                                              container's test  #
+#                based on https://github.com/mli42/containers_test.git	     	 #
 # ****************************************************************************** #
 \e[0m"
 }
@@ -34,55 +35,85 @@ getEmoji () {
 	# 1=integer
 	emoji='';
 	case $1 in
-		0) emoji="${GREEN}OK";;
-		1) emoji="${RED}KO";;
+		0) emoji="${GREEN}✅";;
+		1) emoji="${RED}❌";;
 	esac
 	printf "${BOLD}${emoji}${RESET}"
 }
 
-compare_output () {
-	# 1=diff_file
-	if ! [ -s $1 ]; then
-		return 0
-	fi
-	return 1
-}
-
 printRes () {
-	# 1=file 2=compile 3=bin 4=output
-	printf "%-35s: COMPILE: %s | RET: %s | OUT: %s \n" \
-		"$1" "$(getEmoji $2)" "$(getEmoji $3)" "$(getEmoji $4)"
+	# 1=file 2=result 3=normal fail compilation 4=normal diff
+	# echo "dans print \$1=$1 \$2=$2 \$3=$3 \$4=$4"
+	printf "%-30s: FT <==> STD %s" \
+		"$1" "$(getEmoji $2)"
+	if [ $3 -ne 0 ]
+	then
+		printf " ${DGRAY}=> Normal fail at compilation.${RESET}"
+	fi
+	if [ $4 -eq 1 ]
+	then
+		printf " ${DGRAY}=> Normal diff with maxsize() Mac OS/Linux${RESET}"
+	fi 
+	printf "\n"
 }
 
 isEq () {
 	[ $1 -eq $2 ] && echo 0 || echo 1
 }
 
+compare_diff(){
+	cat $1 | grep "Max_size" &>/dev/null
+	[ "$?" -eq "0" ] && return 1 || return 2;
+}
+
+valid_test()
+{
+	#$1 = $same_output $2=$same_bin $3=$normal_diff
+	if [ $1 -eq 0 ]
+	then
+		if [ $2 -eq 0 ]
+		then
+			return 0
+		fi
+	fi
+	if [ $1 -ne 0 ] && [ $3 -eq 1 ]
+	then
+		return 0
+	fi
+	return 1
+}
+
 test () {
-    logdir="logs" logdiff="diffdir"
-    mkdir -p $logdir $logdiff
+    logdir="logs" logdiff="diffdir" logCompil="logCompil"
+    mkdir -p $logdir $logdiff $logCompil
     container=$(echo $1 | cut -d "/" -f 2)
 	file=$(echo $1 | cut -d "/" -f 3)
 	testname=$(echo $file | cut -d "." -f 1)
 	ft_bin="ft.$container.out"; ft_log="$logdir/ft.$testname.$container.log"
 	std_bin="std.$container.out"; std_log="$logdir/std.$testname.$container.log"
-	std_compile_log="std.$testname.$container.compile.log"
-    ft_compile_log="ft."$testname.$container.compile.log    
+	std_compile_log="$logCompil/std.$testname.$container.compile.log"
+    ft_compile_log="$logCompil/ft."$testname.$container.compile.log    
 
+	#compilation with namespace ft & std
     compile "$1" "ft"  "$ft_bin" $ft_compile_log "ft_";  ft_ret=$?
     compile "$1" "std" "$std_bin" $std_compile_log "std_"; std_ret=$?
 	
-	same_compilation=$(isEq $ft_ret $std_ret)
-	std_compile=$std_ret
-
-	if [ $std_compile -ne 0 ] && \
-		cat $std_compile_log | grep "$container/common.hpp:1" &>/dev/null; then
-		rm -f $std_compile_log
-		printf "${BOLD}${PURPLE}[$container/$testname] Cannot compile${RESET}\n"
-		return;
+	if [ $ft_ret  -eq $std_ret ] && [ $std_ret -ne 0 ]
+	then
+		normal_fail_compil=1	#error expected
+	else
+		normal_fail_compil=0
 	fi
-	rm -f $std_compile_log
-	rm -f $ft_compile_log
+
+	if [ $ft_ret -eq 0 ]
+	then
+		rm -f $std_compile_log	// we rm if OK
+	fi
+
+	if [ $std_ret -eq 0 ]
+	then
+		rm -f $ft_compile_log 	// we rm if OK
+	fi
 
     > $ft_log; > $std_log;
     if [ $ft_ret -eq 0 ]; then
@@ -94,12 +125,22 @@ test () {
 	same_bin=$(isEq $ft_ret $std_ret)
     diff_file="$logdiff/$testname.$container.diff"
 	diff $std_log $ft_log 2>/dev/null 1>"$diff_file";
-    compare_output $diff_file
-    same_output=$?
+
+	if ! [ -s $diff_file ]; then
+		same_output=0
+		normal_diff=0
+	else
+		same_output=1
+		compare_diff $diff_file
+		normal_diff=$?
+	fi
+
+	valid_test $same_output $same_bin $normal_diff
+	ok=$?
 
     rm -f $ft_bin $std_bin
-    printRes "$container/$file" $same_compilation $same_bin $same_output
 
+  	printRes "$container/$file" $ok $normal_fail_compil $normal_diff
     [ -s "$diff_file" ] || rm -f $diff_file $ft_log $std_log &>/dev/null
     rmdir  $deepdir $logdir &>/dev/null
 }
@@ -115,9 +156,8 @@ do_test () {
 }
 
 pheader
-# containers=(vector map stack set)
-containers=(stack)
+containers=(vector map stack set)
 for container in ${containers[@]}; do
-		printf "%40s\n" $container
+		printf "${BOLD}${CYAN}%40s${RESET}\n"  $container
 		do_test $container 2>/dev/null
 done
